% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boot_CI.R
\name{Boot_CI}
\alias{Boot_CI}
\title{Bootstrap Confidence Intervals}
\usage{
Boot_CI(n = 20, alpha = 0.05, pred, resp)
}
\arguments{
\item{n}{A \code{value}(integer) which will be the number of bootstraps run in the function. The default value is 20.}

\item{alpha}{A \code{value}(numeric) that can be between 0 and 1, but is usually closer to 0. The default value is 0.05.}

\item{pred}{A \code{matrix}(numeric) containing variables that will be used to predict the response.}

\item{resp}{A \code{vector}(numeric) containing values that represent a response variable.}
}
\value{
A \code{matrix}(numeric) where the number of rows is equal to the amount of betas and 2 columns. Each row represents a confidence interval for each predictor at the chosen confidence level.
}
\description{
This function will conduct a bootstrap routine for each predictor or beta value for the logistic regression curve. The user is able to choose the amount of bootstraps to conduct as well as the confidence level for the intervals.
}
\examples{
mydata <- read.csv("expenses.csv")
mydata$sex <- factor(mydata$sex, levels = c("male","female"), labels = c(0,1))
mydata$smoker <- factor(mydata$smoker, levels = c("yes","no"), labels = c(1,0))
mydata$region_northeast <- ifelse(mydata$region=="northeast",1,0)
mydata$region_northwest <- ifelse(mydata$region=="northwest",1,0)
mydata$region_southeast <- ifelse(mydata$region=="southeast",1,0)
mydata$region_northeast <- ifelse(mydata$region=="northeast",1,0)
mydata$region_northwest <- ifelse(mydata$region=="northwest",1,0)
mydata$region_southeast <- ifelse(mydata$region=="southeast",1,0)
mydata <- mydata[,-6]
mydata <- as.matrix(mydata)
mydata <- as.numeric(mydata)
mydata <- matrix(mydata, ncol=9)
X <- mydata[, c(1:4,6:9)]
Y <- mydata[, 5]
Boot_CI_demo <- Boot_CI(n=20, alpha=0.05, pred= X, resp= Y)
}
\author{
Yizhong Zhang, Alexander Miller
}
