% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{crossentropy}
\alias{crossentropy}
\title{binary cross entropy}
\usage{
crossentropy(y, p)
}
\arguments{
\item{y}{true label (0/1)}

\item{p}{predicted probability}
}
\description{
binary cross entropy
}
\note{
\code{-(y * log(p) + (1 - y) * log (1 - p))}
}
