% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_fisherScoring.R
\name{fishScore}
\alias{fishScore}
\title{The Fisher Scoring algorithm}
\usage{
fishScore(x, y, maxIter = 25, precision = 1e-08)
}
\arguments{
\item{x}{Matrix/vector containing the design matrix (including intercept).}

\item{y}{Matrix/vector containing the response variable.}

\item{maxIter}{Maximum number of iterations of the Fisher scoring
algorithm.}

\item{precision}{Degree of precision of the Fisher scoring algorithm.}
}
\value{
A list with maximum likelihood estimation results for further computation.
}
\description{
Fisher scoring is an optimization algorithm used to compute the
paramenters for a logistic regression model.
}
\details{
The algorithm is applied in the \code{\link{logMod}} regression estimation
function.
}
\examples{
set.seed(1)
x1 <- rnorm(100, mean = 10);
x2 <- rnorm(100, mean = 10);
x0 <- c(rep(1, times = 100));
someVariable <- x1 - x2
x <- cbind(x0, x1, x2)
y <- rbinom(100, 1, exp(someVariable) / (1 + exp(someVariable)))
fit <- fishScore(x = x, y = y)

}
\references{
Hastie, T. J. and Pregibon, D. (1992): Generalized linear models.
Statistical Models in S, J. M. Chambers and T. J. Hastie (eds), USA:
Wadsworth & Brooks/Cole.
}
