% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_logistic}
\alias{fit_logistic}
\title{fit a logistic regression model}
\usage{
fit_logistic(x, y, learning_rate = 0.05, n_epochs = 1)
}
\arguments{
\item{x}{predictor matrix (n x p)}

\item{y}{response matrix (n x 1)}

\item{learning_rate}{learning rate}

\item{n_epochs}{number of epoches}
}
\description{
fit a logistic regression model
}
