% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_logistic_ridge}
\alias{fit_logistic_ridge}
\title{fit a regularized logistic regression model (ridge penalty)}
\usage{
fit_logistic_ridge(x, y, learning_rate = 0.05, n_epochs = 1, lambda = 1)
}
\arguments{
\item{x}{predictor matrix (n x p)}

\item{y}{response matrix (n x 1)}

\item{learning_rate}{learning rate}

\item{n_epochs}{number of epoches}

\item{lambda}{regularization parameter}
}
\description{
fit a regularized logistic regression model (ridge penalty)
}
