% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Starting Values for Beta}
\usage{
init(resp, pred)
}
\arguments{
\item{resp}{A \code{vector}(numeric) containing values that represent a response variable.}

\item{pred}{A \code{matrix}(numeric) containing variables that will be used to predict the response.}
}
\value{
A \code{vector} containing beta's or coefficients for each variable in the pred matrix that predict the resp vector.
}
\description{
This function creates a vector to be the starting point for optimization using the least squares formula.
}
\examples{
mydata <- read.csv("expenses.csv")
mydata$sex <- factor(mydata$sex, levels = c("male","female"), labels = c(0,1))
mydata$smoker <- factor(mydata$smoker, levels = c("yes","no"), labels = c(1,0))
mydata$region_northeast <- ifelse(mydata$region=="northeast",1,0)
mydata$region_northwest <- ifelse(mydata$region=="northwest",1,0)
mydata$region_southeast <- ifelse(mydata$region=="southeast",1,0)
mydata$region_northeast <- ifelse(mydata$region=="northeast",1,0)
mydata$region_northwest <- ifelse(mydata$region=="northwest",1,0)
mydata$region_southeast <- ifelse(mydata$region=="southeast",1,0)
mydata <- mydata[,-6]
mydata <- as.matrix(mydata)
mydata <- as.numeric(mydata)
mydata <- matrix(mydata, ncol=9)
X <- mydata[, c(1:4,6:9)]
Y <- mydata[, 5]
init_beta_demo <- init(resp = Y, pred= X)
}
\author{
Yizhong Zhang, Alexander Miller
}
