% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer.R
\name{initialize_glorot_normal}
\alias{initialize_glorot_normal}
\title{Glorot/Xavier normal initializer}
\usage{
initialize_glorot_normal(fan_in, fan_out, seed = NULL)
}
\arguments{
\item{fan_in}{number of input elements}

\item{fan_out}{number of output elements}

\item{seed}{random number seed}
}
\description{
Glorot/Xavier normal initializer
}
