% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_plot.R
\name{log_plot}
\alias{log_plot}
\title{Plot of the Logistic Curve}
\usage{
log_plot(pred, resp)
}
\arguments{
\item{pred}{A \code{matrix}(numeric) containing variables that will be used to predict the response.}

\item{resp}{A \code{vector}(numeric) containing values that represent a response variable.}
}
\value{
A \code{plot} that shows the fitted logistic curve to the responses, where the y-axis is the binary response y while the x-axis represents a sequence of values from the range of fitted values from the logistic regression.
}
\description{
This function will plot the optimized logistic curve for the given data.
}
\examples{
mydata <- read.csv("expenses.csv")
mydata$sex <- factor(mydata$sex, levels = c("male","female"), labels = c(0,1))
mydata$smoker <- factor(mydata$smoker, levels = c("yes","no"), labels = c(1,0))
mydata$region_northeast <- ifelse(mydata$region=="northeast",1,0)
mydata$region_northwest <- ifelse(mydata$region=="northwest",1,0)
mydata$region_southeast <- ifelse(mydata$region=="southeast",1,0)
mydata$region_northeast <- ifelse(mydata$region=="northeast",1,0)
mydata$region_northwest <- ifelse(mydata$region=="northwest",1,0)
mydata$region_southeast <- ifelse(mydata$region=="southeast",1,0)
mydata <- mydata[,-6]
mydata <- as.matrix(mydata)
mydata <- as.numeric(mydata)
mydata <- matrix(mydata, ncol=9)
X <- mydata[, c(1:4,6:9)]
Y <- mydata[, 5]
log_plot(pred= X, resp= Y)
}
\author{
Yizhong Zhang, Alexander Miller
}
