% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_logMod_S3.R
\name{logMod}
\alias{logMod}
\title{Estimating logistic regression model}
\usage{
logMod(formula, data = list(), ...)
}
\arguments{
\item{formula}{an object of class "formula": represents the
model to be fitted.}

\item{data}{an optional data frame, list or environment containing the
variables in the model. If the variables are not found in \code{data}, the
variables are taken from \code{environment(formula)}.}

\item{...}{optional specifications used for the setup of the Fisher scoring
algorithm (see \code{\link{fishScore}}).}
}
\value{
\code{logMod} returns an object of \code{\link[base]{class}}
    \emph{logMod}.
}
\description{
\code{logMod} is used to fit logistic regression models with a binaray
response variable.
}
\examples{
set.seed(1)
x1 <- rnorm(100, mean = 10);
x2 <- rnorm(100, mean = 10);
someVariable <- x1 - x2
x <- cbind(x1, x2)
y <- rbinom(100, 1, exp(someVariable) / (1 + exp(someVariable)))
fit <- logMod(y ~ x)

}
