% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg.R
\name{logreg}
\alias{logreg}
\title{Format data and perform binomial logistic regression in a single command}
\usage{
logreg(
  data,
  targets,
  response,
  binomial_target,
  binomial_alternate = NULL,
  mode = "binomial",
  splits,
  partition = NULL,
  endpoint = NULL
)
}
\arguments{
\item{data}{The dataset which stores the explanatory and response variables.}

\item{targets}{A vector of each explanatory variable. Must be in quotes.}

\item{response}{The response variable. Must be in quotes.}

\item{binomial_target}{The target value of the response variable for binomial regression}

\item{binomial_alternate}{(optional) The secondary value of the response variable for binomial regression (if unused, the default is "other")}

\item{mode}{(optional) What type of regression you want to do (the only currently supported type is "binomial")}

\item{splits}{The number of categories you want your explanatory variable split up into (higher values are more accurate but may not work, I recommend 10)}

\item{partition}{(optional) Set to TRUE to cut your dataset down to a managable size for modelling, set to FALSE to bypass this. If you bypass this, it may crash R.}

\item{endpoint}{What you want as the output of this command. Options: "check" (check data compatability), format (format response variable), split (split explanatory variables), numbers (view the regression model as text), or model (create a regression model and ggplot2 visualization)}
}
\description{
Format data and perform binomial logistic regression in a single command
}
