% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_value.R
\name{loss_value}
\alias{loss_value}
\title{Loss Function}
\usage{
loss_value(pred, resp, beta)
}
\arguments{
\item{pred}{A \code{matrix}(numeric) containing variables that will be used to predict the response.}

\item{resp}{A \code{vector}(numeric) containing values that represent a response variable.}

\item{beta}{A \code{vector}(numeric) containing beta's or coefficients for each variable in the pred matrix that predict the resp vector.}
}
\value{
A \code{value}(numeric) computed by this loss function.
}
\description{
This is the "loss function" that can be used to optimize the chosen beta's based on the value given by this function. This will be used for logistic regression.
}
\author{
Yizhong Zhang, Alexander Miller
}
