% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_plot_S3.R
\name{plot.logMod}
\alias{plot.logMod}
\title{S3 Plot Method for logMod Objects}
\usage{
\method{plot}{logMod}(x, ...)
}
\arguments{
\item{x}{an object of class logMod.}

\item{...}{methods are required to include (at least) the same arguments
as their generic functions.}
}
\description{
An internal function, which defines a \code{\link{plot}} method for an object
of class \code{logMod}.
}
