% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metric.R
\name{plot_metric}
\alias{plot_metric}
\title{Metric Plot}
\usage{
plot_metric(resp, pred, metric)
}
\arguments{
\item{resp}{A \code{vector}(numeric) containing values that represent a response variable.}

\item{pred}{A \code{matrix}(numeric) containing variables that will be used to predict the response.}

\item{metric}{This can be any of the following metrics: "Prevalence", "Accuracy","Sensitivity", "Specificity", "False Discovery Rate","Diagnostic Odds Ratio"}
}
\value{
A \code{plot} that shows
}
\description{
This function conducts logistic regression on data and plots a given metric that visualizes the performance of the regression.
}
\examples{
mydata <- read.csv("expenses.csv")
mydata$sex <- factor(mydata$sex, levels = c("male","female"), labels = c(0,1))
mydata$smoker <- factor(mydata$smoker, levels = c("yes","no"), labels = c(1,0))
mydata$region_northeast <- ifelse(mydata$region=="northeast",1,0)
mydata$region_northwest <- ifelse(mydata$region=="northwest",1,0)
mydata$region_southeast <- ifelse(mydata$region=="southeast",1,0)
mydata$region_northeast <- ifelse(mydata$region=="northeast",1,0)
mydata$region_northwest <- ifelse(mydata$region=="northwest",1,0)
mydata$region_southeast <- ifelse(mydata$region=="southeast",1,0)
mydata <- mydata[,-6]
mydata <- as.matrix(mydata)
mydata <- as.numeric(mydata)
mydata <- matrix(mydata, ncol=9)
X <- mydata[, c(1:4,6:9)]
Y <- mydata[, 5]
plot_metric(resp=Y, pred=X, metric = "Prevalence")
plot_metric(resp=Y, pred=X, metric = "Accuracy")
}
\author{
Yizhong Zhang, Alexander Miller
}
