% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.logreg}
\alias{predict.logreg}
\title{make predictions from a fitted model}
\usage{
\method{predict}{logreg}(object, newdata, ...)
}
\arguments{
\item{object}{model object}

\item{newdata}{new predictor matrix}

\item{...}{unused}
}
\description{
make predictions from a fitted model
}
