% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_logMod_S3.R
\name{print.logMod}
\alias{print.logMod}
\title{S3 Print Method for logMod Objects}
\usage{
\method{print}{logMod}(x, ...)
}
\arguments{
\item{x}{a logMod object.}

\item{...}{optional arguments to \code{print} methods.}
}
\value{
a short summary of the estimation results.
}
\description{
An internal function, which defines a \code{\link{print}} method for an
object of class \code{logMod}.
}
