% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_summary_S3.R
\name{print.summary.logMod}
\alias{print.summary.logMod}
\title{S3 Print Method for summary.logMod Objects}
\usage{
\method{print}{summary.logMod}(x, ...)
}
\arguments{
\item{x}{an object of class "summary.logMod".}

\item{...}{optional arguments to \code{summary} methods.}
}
\value{
a summary of the estimation results.
}
\description{
An internal function, which defines a \code{\link{print}} method for an
object of class \code{\link{summary.logMod}}.
}
