% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_apps.R
\name{run_apps}
\alias{run_apps}
\title{Shiny Apps Loader}
\usage{
run_apps(
  app_name = NULL,
  theme = "flatly",
  width = "100\%",
  height = "800px",
  more_opts = list(NA),
  launch.browser = TRUE,
  ...
)
}
\description{
This will run the shiny applications in this project. \cr
             As a reminder, the applications should be in quotes.  Example: run_apps("logreg_gui") \cr
             See details for descriptions of the various apps.
}
\details{
Application: "logreg_gui". \cr
         Logreg_gui is a step-wise logistic regression model.  A user may choose various variables to include in the model and the app will provide you with the G-Statistic for each response.  The variable with the highest G-Statistic will provide the most information for your model. \cr
         At this time, the structure of the data requires that the response be in column (1) with the column name "Transition".
}
