% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_summary_S3.R
\name{summary.logMod}
\alias{summary.logMod}
\title{S3 Summary Method for logMod Objects}
\usage{
\method{summary}{logMod}(x, ...)
}
\arguments{
\item{x}{a logMod object for which a summary should be returned.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
a list with statistics summarazing the model estimation results.
}
\description{
An internal function, which defines a \code{\link{summary}} method for
an object of class \code{logMod}. The output of this function gives further
statistics that inform about the quality of the estimation results and
that might be used for model comparison.
}
