% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Test_Data}
\alias{Test_Data}
\title{Sample Nation-state Statistics}
\format{A \code{tibble} with 60 rows and 6 variables
\tabular{rlll}{
[,1] \tab Transition \tab Binary markov chain indicator for transitioning to conflic \tab \bold{numeric}\cr
[,2] \tab GDP \tab Gross Domestic Product \tab \bold{numeric}\cr
[,3] \tab IMR \tab Infant Mortality Rate \tab \bold{numeric}\cr
[,4] \tab PD \tab Population Density \tab \bold{numeric}\cr
[,5] \tab FT2Y \tab Two Year Freedom Trend \tab \bold{numeric}\cr
[,6] \tab FT5Y \tab Five Year Freedom Trend \tab \bold{numeric}
}}
\description{
A dataset containing a sample of statistics from Arab nation states between 2004 and 2009.
}
\concept{data-testing}
